pyproject.toml
[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[project]
name = "gxucnm"
version = "0.1.0"
description = "GXU Campus Network Manager"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["requests>=2.32.3", "python-dotenv>=1.1.0", "ruff>=0.15.10"]

[project.scripts]
gxucnm = "gxucnm.cli:main"

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.setuptools.packages.find]
where = ["src"]
